+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   * Scope Module
   * Sweeper Module
   * DAQ Module
-* Renamed `zhinst.toolkit.nodetree.nodes.WildcardResult` to `zhinst.toolkit.nodetree.helpers.NodeDict`
+* Renamed `zhinst.toolkit.nodetree.node.WildcardResult` to `zhinst.toolkit.nodetree.helpers.NodeDict`
 * Added `active_validation` argument to `CommandTable`. By disabling it, `CommandTable` does not actively 
   validate the inputs and therefore it improves the speed for command table creation.
 * Adapt `awg.enable_sequencer` to check the acknowledged value instead of using `wait_for_state_change`. This makes it much more stable when used with short sequences.